deploy: add @AuthenticationPrincipal on Controller Codes

+++ src/main/kotlin/com/bandchu/api/domain/posts/service/CommentService.kt
@@ -2,6 +2,8 @@ package com.bandchu.api.domain.posts.service
 
 import com.bandchu.api.domain.posts.dto.response.CommentResponse
 import com.bandchu.api.domain.posts.repository.CommentRepository
+import com.bandchu.api.global.exception.BusinessException
+import com.bandchu.api.global.exception.ErrorCode
 import org.springframework.stereotype.Service
 
 @Service
@@ -19,21 +21,29 @@ class CommentService(
     }
 
     //댓글 삭제
-    fun deleteComment(commentId: Long): Long {
-//        if (commentRepository.findByCommentId(commentId) != null){
-//            throw BusinessException(ErrorCode.)
-//        }
+    fun deleteComment(memberId: Long, commentId: Long): Long {
+        if (commentRepository.findByCommentId(commentId) == null){
+                throw BusinessException(ErrorCode.COMMENT_NOT_FOUND)
+        }
+
+        //댓글 작성자가 아니면
+        if ( commentRepository.findByCommentId(commentId)!!.memberId != memberId)
+            throw BusinessException(ErrorCode.COMMENT_FORBIDDEN)
 
         commentRepository.deleteComment(commentId)
         return commentId
     }
 
     //댓글 업데이트
-    fun updateComment(commentId: Long, content: String): CommentResponse {
-        //        if (commentRepository.findByCommentId(commentId) != null){
-//            throw BusinessException(ErrorCode.)
-//        }
+    fun updateComment(memberId: Long, commentId: Long, content: String): CommentResponse {
+        if (commentRepository.findByCommentId(commentId) == null)
+            throw BusinessException(ErrorCode.COMMENT_NOT_FOUND)
+
+        //댓글 작성자가 아니면
+        if (commentRepository.findByCommentId(commentId)!!.memberId != memberId)
+            throw BusinessException(ErrorCode.COMMENT_FORBIDDEN)
 
         return commentRepository.updateComment(commentId, content)
     }
+
 }
